Review Section Frontend Cleanup
