+++ MTM-Frontend/src/components/Form/Theme.tsx
@@ -3,8 +3,11 @@ import { createTheme } from "@mui/material/styles";
 const PrimaryMainTheme = createTheme({
     palette: {
         primary: {
-            main: "#A4A4A4"
-        }
+            main: "#A4A4A4",
+        },
+        secondary: {
+            main: "#FFFFFF",
+        },
     }
 });
